pyproject.toml
[tool.poetry]
name = "ml-playground-stream"
version = "0.1.0"
description = ""
authors = ["yuya-okada527 <yuya.okada527@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
google-cloud-pubsub = "^2.4.1"
google-cloud-storage = "^1.37.1"
more-itertools = "^8.7.0"
pydantic = "^1.8.1"
python-dotenv = "^0.17.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
